hv_functions had voset value as int, changed to float

+++ hv_c_lib/hv_functions.c
@@ -90,7 +90,7 @@ float get_ch_vmon(int sys_handle, int slot, int chan) {
 }
 
 
-void set_ch_v0(int sys_handle, int slot, int chan, int value) {
+void set_ch_v0(int sys_handle, int slot, int chan, float value) {
 	unsigned short ch_list[] = { chan };
 	unsigned short ch_value[] = { value };
 	CAENHVRESULT ret = CAENHV_SetChParam(sys_handle, slot, "V0Set", 1, ch_list, ch_value);
